+++ src/curvesimulator/__init__.py
@@ -3,7 +3,7 @@ from .curvesim import CurveSimulator
 from .cs_animation import CurveSimAnimation
 from .cs_bodies import CurveSimBodies
 from .cs_body import CurveSimBody
-from .cs_flux_data import csv2df
+from .cs_flux_data import CurveSimFluxData, csv2df
 from .cs_lightcurve import CurveSimLightcurve
 from .cs_mcmc import CurveSimMCMC
 from .cs_parameters import CurveSimParameters, FittingParameter